[BE] refactor: 북마크 필터 패키지 이동 #643
- BookmarkFilter 관련 클래스들을 적합한 dto 하위 패키지로 이동
- StringToBookmarkFilterConverter 클래스 이름을 BookmarkFilterConverter로 변경
