+++ backend/app/src/main/java/com/onair/hearit/app/bookmark/dto/param/BookmarkFilter.java
@@ -1,7 +1,11 @@
-package com.onair.hearit.app.bookmark;
+package com.onair.hearit.app.bookmark.dto.param;
 
 import java.util.Arrays;
+import lombok.Getter;
+import lombok.RequiredArgsConstructor;
 
+@Getter
+@RequiredArgsConstructor
 public enum BookmarkFilter {
 
     UNFINISHED("unfinished"),
@@ -10,18 +14,10 @@ public enum BookmarkFilter {
 
     private final String name;
 
-    BookmarkFilter(String name) {
-        this.name = name;
-    }
-
     public static BookmarkFilter fromName(String queryParam) {
         return Arrays.stream(values())
                 .filter(value -> value.name.equalsIgnoreCase(queryParam))
                 .findFirst()
                 .orElseThrow(IllegalArgumentException::new);
-    }
-
-    public String getName() {
-        return name;
     }
 }
