+++ backend/app/src/main/java/com/onair/hearit/app/common/config/ConverterWebConfig.java
@@ -1,5 +1,6 @@
-package com.onair.hearit.app.bookmark;
+package com.onair.hearit.app.common.config;
 
+import com.onair.hearit.app.bookmark.dto.converter.BookmarkFilterConverter;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.format.FormatterRegistry;
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
@@ -9,6 +10,6 @@ public class ConverterWebConfig implements WebMvcConfigurer {
 
     @Override
     public void addFormatters(FormatterRegistry registry) {
-        registry.addConverter(new StringToBookmarkFilterConverter());
+        registry.addConverter(new BookmarkFilterConverter());
     }
 }
