+++ backend/app/src/test/java/com/onair/hearit/app/bookmark/dto/converter/BookmarkFilterConverterTest.java
@@ -1,14 +1,15 @@
-package com.onair.hearit.app.bookmark;
+package com.onair.hearit.app.bookmark.dto.converter;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
+import com.onair.hearit.app.bookmark.dto.param.BookmarkFilter;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
 
-class StringToBookmarkFilterConverterTest {
+class BookmarkFilterConverterTest {
 
-    private StringToBookmarkFilterConverter converter = new StringToBookmarkFilterConverter();
+    private BookmarkFilterConverter converter = new BookmarkFilterConverter();
 
     @ParameterizedTest
     @CsvSource({
